+++ src/Fauna/create_herbivore_cohort.cpp
@@ -6,6 +6,7 @@
  */
 #include "create_herbivore_cohort.h"
 #include "herbivore_cohort.h"
+#include "parameters.h"
 
 using namespace Fauna;
 
@@ -24,9 +25,10 @@ HerbivoreCohort CreateHerbivoreCohort::operator()(const double ind_per_km2,
   const int age_days = age_years * 365;
   if (age_days == 0)
     // Call birth constructor
-    return HerbivoreCohort(&get_hft(), sex, ind_per_km2);
+    return HerbivoreCohort(&get_hft(), sex, ind_per_km2,
+                           get_params().metabolizable_energy);
   else
     // Call establishment constructor
     return HerbivoreCohort(age_days, get_body_condition(age_days), &get_hft(),
-                           sex, ind_per_km2);
+                           sex, ind_per_km2, get_params().metabolizable_energy);
 }