+++ src/Fauna/herbivore_cohort.h
@@ -34,33 +34,30 @@ class HerbivoreCohort : public HerbivoreBase {
    * of physiological maximum [kg/kg].
    * \param hft Herbivore functional type.
    * \param sex The sex of the herbivore.
+   * \param metabolizable_energy The (constant) metabolizable energy content
+   * for the forage types [MJ/kgDM]. See: \ref Parameters::metabolizable_energy
    * \throw std::invalid_argument if any parameter is invalid
    *
    * \param ind_per_km2 Initial individual density [ind/km²].
    * Can be 0.0, but must not be negative.
    */
   HerbivoreCohort(const int age_days, const double body_condition,
-                  const Hft* hft, const Sex sex, const double ind_per_km2);
+                  const Hft* hft, const Sex sex, const double ind_per_km2,
+                  const ForageEnergyContent& metabolizable_energy);
 
   /// Birth constructor
   /**
    * \param hft Herbivore functional type.
    * \param sex The sex of the herbivore.
+   * \param metabolizable_energy The (constant) metabolizable energy content
+   * for the forage types [MJ/kgDM]. See: \ref Parameters::metabolizable_energy
    * \throw std::invalid_argument if any parameter is invalid
    *
    * \param ind_per_km2 Initial individual density [ind/km²].
    * Can be 0.0, but must not be negative.
    */
-  HerbivoreCohort(const Hft* hft, const Sex sex, const double ind_per_km2);
-
-  /// Copy constructor.
-  HerbivoreCohort(const HerbivoreCohort& other);
-
-  /// Copy assignment operator.
-  HerbivoreCohort& operator=(const HerbivoreCohort& other);
-
-  /// Destructor.
-  ~HerbivoreCohort(){};
+  HerbivoreCohort(const Hft* hft, const Sex sex, const double ind_per_km2,
+                  const ForageEnergyContent& metabolizable_energy);
 
   /// Check if this and the other cohort are of the same age
   /**
@@ -81,9 +78,9 @@ class HerbivoreCohort : public HerbivoreBase {
    * \param other The other cohort that is merged into `this`.
    * The density of `other` will be reduced.
    * \throw std::invalid_argument If `fraction` not in [0,1].
-   * \throw std::logic_error If the other cohort is not
-   * compatible: different age, different HFT, or different
-   * sex.
+   * \throw std::logic_error If the other cohort is not compatible: different
+   * age, different HFT, different metabolizable_energy, or different sex. See
+   * \ref HerbivoreBase::constant_members_match()
    */
   void merge(HerbivoreCohort& other);