+++ src/Fauna/parameters.cpp
@@ -30,6 +30,13 @@ bool Parameters::is_valid(std::string& messages) const {
       is_valid = false;
     }
 
+  for (const auto ft : FORAGE_TYPES)
+    if (metabolizable_energy[ft] == 0.0)
+      stream << "metabolizable_energy." << get_forage_type_name(ft)
+             << " is zero. "
+             << "This forage type will effectively be deactivated."
+             << std::endl;
+
   //------------------------------------------------------------
 
   // convert stream to string